refactored saving, works with normal function, to be adjusted to buffermanager

+++ util/GlobalAddressHash.h
@@ -17,7 +17,7 @@ template<class T> class GlobalAddressHash;
         uint64_t operator()(const defs::SendGlobalAddr &ga) const
         {
             uint64_t h1 = std::hash<size_t>()(ga.size);
-            uint64_t h2 = std::hash<uint64_t>()(ga.ptr);
+            uint64_t h2 = std::hash<uintptr_t>()(ga.ptr);
             uint64_t h3 = std::hash<uint64_t>()(ga.id);
             uint64_t h4 = std::hash<bool>()(ga.isFile);
             return (h1 ^ (h2 ^(h3 ^  (h4<<1))));
